requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.3
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==15.1.3
idna==3.4
joblib==1.2.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.4
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
spotipy==2.21.0
sqlparse==0.4.3
threadpoolctl==3.1.0
toposort==1.7
tzdata==2022.6
urllib3==1.26.12
wrapt==1.14.1
